+++ src/Vulkan/RayTracing/RayTracingPipeline.hpp
@@ -44,7 +44,7 @@ namespace Vulkan::RayTracing
 		uint32_t TriangleHitGroupIndex() const { return triangleHitGroupIndex_; }
 		uint32_t ProceduralHitGroupIndex() const { return proceduralHitGroupIndex_; }
 
-		VkDescriptorSet DescriptorSet(uint32_t index) const;
+		VkDescriptorSet DescriptorSet(size_t index) const;
 		const class PipelineLayout& PipelineLayout() const { return *pipelineLayout_; }
 
 	private:
